LayoutTests/media/video-seek-with-negative-playback-expected.txt

EVENT(canplaythrough)
EVENT(seeked)
EVENT(timeupdate)
EVENT(timeupdate)
EXPECTED (video.currentTime < '3') OK
END OF TEST
